Revert "Show errors on file upload form, format data to be validated by api"
