+++ exporter/applications/helpers/date_fields.py
@@ -1,5 +1,15 @@
 import re
-from core.helpers import format_date
+
+
+def format_date(data, date_field):
+    year = data.get(date_field + "year", "")
+    month = data.get(date_field + "month", "")
+    if len(month) == 1:
+        month = "0" + month
+    day = data.get(date_field + "day", "")
+    if len(day) == 1:
+        day = "0" + day
+    return f"{year}-{month}-{day}"
 
 
 def format_date_fields(data):
