+++ exporter/goods/services.py
@@ -1,8 +1,8 @@
 from http import HTTPStatus
 
+from exporter.applications.helpers.date_fields import format_date
 from core import client
-from core.helpers import convert_parameters_to_query_params, format_date
-from exporter.goods.helpers import serialize_goods_data, add_firearm_details_to_data
+from core.helpers import convert_parameters_to_query_params
 
 
 def get_goods(
@@ -23,11 +23,137 @@ def get_good_details(request, pk):
 
 
 def post_goods(request, json):
-    serialize_goods_data(request, json)
+    json["item_category"] = json.get("item_category", "group2_firearms")
+
+    if "is_pv_graded" in json and json["is_pv_graded"] == "yes":
+        if "reference" in json:
+            json["pv_grading_details"] = {
+                "grading": json["grading"],
+                "custom_grading": json["custom_grading"],
+                "prefix": json["prefix"],
+                "suffix": json["suffix"],
+                "issuing_authority": json["issuing_authority"],
+                "reference": json["reference"],
+                "date_of_issue": format_date(json, "date_of_issue"),
+            }
+
+    if "item_category" in json and json["item_category"] == "group2_firearms":
+        add_firearm_details_to_data(json)
 
     data = client.post(request, "/goods/", json)
 
+    if data.status_code == HTTPStatus.OK:
+        data.json().get("good"), data.status_code
     return data.json(), data.status_code
+
+
+def add_section_certificate_details(firearm_details, json):
+    if "section_certificate_step" in json:
+        firearm_details["is_covered_by_firearm_act_section_one_two_or_five"] = json.get(
+            "is_covered_by_firearm_act_section_one_two_or_five", ""
+        )
+        firearm_details["firearms_act_section"] = json.get("firearms_act_section", "")
+    if "firearms_certificate_uploaded" in json:
+        certificate_missing = json.get("section_certificate_missing", False)
+        if not certificate_missing:
+            firearm_details["section_certificate_number"] = json.get("section_certificate_number")
+            formatted_section_certificate_date = format_date(json, "section_certificate_date_of_expiry")
+            firearm_details["section_certificate_date_of_expiry"] = (
+                formatted_section_certificate_date if formatted_section_certificate_date != "--" else None
+            )
+            firearm_details["section_certificate_missing"] = False
+            firearm_details["section_certificate_missing_reason"] = ""
+        else:
+            firearm_details["section_certificate_missing"] = True
+            firearm_details["section_certificate_missing_reason"] = json.get("section_certificate_missing_reason", "")
+            firearm_details["section_certificate_number"] = ""
+
+    return firearm_details
+
+
+def add_identification_marking_details(firearm_details, json):
+    if "number_of_items_step" in json:
+        try:
+            firearm_details["number_of_items"] = int(json.get("number_of_items"))
+        except ValueError:
+            firearm_details["number_of_items"] = 0
+
+    if "identification_markings_step" in json:
+        # parent component doesnt get sent when empty unlike the remaining form fields
+        firearm_details["has_identification_markings"] = json.get("has_identification_markings", "")
+        firearm_details["no_identification_markings_details"] = json.get("no_identification_markings_details")
+        del json["no_identification_markings_details"]
+
+    if "capture_serial_numbers_step" in json:
+        try:
+            number_of_items = int(json.get("number_of_items"))
+        except ValueError:
+            number_of_items = 0
+
+        serial_numbers = []
+        for i in range(number_of_items):
+            serial_numbers.append(json.get(f"serial_number_input_{i}", ""))
+        firearm_details["serial_numbers"] = serial_numbers
+
+    return firearm_details
+
+
+def add_firearm_details_to_data(json):
+    """
+    Return a firearm_details dictionary to be used when creating/editing a group 2 firearm good
+    Mutable - items in firearm_details are removed from the original json (duplicates)
+    """
+    firearm_details = {}
+    if "product_type_step" in json:
+        # parent component doesnt get sent when empty unlike the remaining form fields
+        firearm_details["type"] = json.get("type")
+
+    if "sporting_shotgun_step" in json:
+        firearm_details["type"] = json.get("type")
+        firearm_details["is_sporting_shotgun"] = json.get("is_sporting_shotgun")
+    elif firearm_details and "is_sporting_shotgun" not in firearm_details:
+        firearm_details["is_sporting_shotgun"] = False
+
+    firearm_details = add_identification_marking_details(firearm_details, json)
+
+    if "firearm_year_of_manufacture_step" in json:
+        firearms_year_of_manufacture = json.pop("year_of_manufacture")
+        if firearms_year_of_manufacture == "":
+            firearms_year_of_manufacture = None
+        firearm_details["year_of_manufacture"] = firearms_year_of_manufacture
+    elif firearm_details and "year_of_manufacture" not in firearm_details:
+        firearm_details["year_of_manufacture"] = 0
+
+    if "is_replica_step" in json:
+        firearm_details["type"] = json.get("type")
+        firearm_details["is_replica"] = json.get("is_replica")
+        firearm_details["replica_description"] = json.get("replica_description", "")
+        del json["replica_description"]
+
+    if "firearm_calibre_step" in json:
+        firearm_calibre = json.pop("calibre")
+        if firearm_calibre == "":
+            firearm_calibre = None
+        firearm_details["calibre"] = firearm_calibre
+
+    firearm_details = add_section_certificate_details(firearm_details, json)
+
+    for name in [
+        "date_of_deactivation",
+        "has_proof_mark",
+        "no_proof_mark_details",
+        "is_deactivated",
+        "deactivation_standard",
+        "deactivation_standard_other",
+        "is_deactivated_to_standard",
+    ]:
+        if name in json:
+            firearm_details[name] = json.pop(name)
+
+    if firearm_details:
+        json["firearm_details"] = firearm_details
+
+    return json
 
 
 def validate_good(request, json):
@@ -48,7 +174,7 @@ def edit_good_details(request, pk, json):
 
 
 def edit_good_firearm_details(request, pk, json):
-    add_firearm_details_to_data(request, json)
+    add_firearm_details_to_data(json)
     return edit_good_details(request, pk, json)
